+++ src/engine/source/builder/builders/opBuilderHelperMap.hpp
@@ -229,6 +229,20 @@ base::Expression opBuilderHelperIPVersionFromIPStr(const std::any& definition);
  */
 base::Expression opBuilderHelperEpochTimeFromSystem(const std::any& definition);
 
+//*************************************************
+//*              Checksum and hash                *
+//*************************************************
+
+/**
+ * @brief Builds helper SHA1 hash calculated from a strings or a reference.
+ * <field>: +hash_sha1/<string1>|$<string_reference1>
+ *
+ * @param definition Definition of the operation to be built.
+ * @return base::Expression The Lifter with the SHA1 hash.
+ * @throw std::runtime_error if the parameter size is not one.
+ */
+base::Expression opBuilderHelperHashSHA1(const std::any& definition);
+
 } // namespace builder::internals::builders
 
 #endif // _OP_BUILDER_HELPER_MAP_H
